refactor(knockdog): bottom nav bar height 상수 사용, layout 수정
