+++ apps/knockdog/app/layout.tsx
@@ -2,11 +2,10 @@ import { cn } from '@knockdog/ui/lib';
 
 import { suit } from './font';
 import './globals.css';
-
 import { NaverMapProvider } from '@knockdog/naver-map';
 import { ReactQueryProvider } from '@app/providers/ReactQueryProvider';
 import { HeaderProvider, HeaderWrapper } from '@widgets/Header';
-import { BottomNavigationBar } from '@widgets/BottomNavigationBar';
+import { BottomNavigationBar } from '@widgets/bottom-bar';
 
 export default function RootLayout({
   children,
@@ -21,10 +20,10 @@ export default function RootLayout({
             <HeaderProvider>
               <div className='flex h-dvh flex-col'>
                 <HeaderWrapper />
-                <div className='flex-1'>
-                  <div>{children}</div>
+                {children}
+                <div className='fixed inset-x-0 bottom-0 z-50'>
+                  <BottomNavigationBar />
                 </div>
-                <BottomNavigationBar />
               </div>
             </HeaderProvider>
           </NaverMapProvider>
