+++ apps/knockdog/src/shared/constants/ui.ts
@@ -1,0 +1,1 @@
+export const BOTTOM_BAR_HEIGHT = 68;
