+++ apps/knockdog/src/widgets/bottom-bar/index.ts
